+++ krml/cli.py
@@ -7,6 +7,7 @@ from rdflib import ConjunctiveGraph
 from rdflib.plugin import PluginException
 from rich import print_json
 from rich.console import Console
+from rich.syntax import Syntax
 from typing_extensions import Annotated
 
 from krml import Template
@@ -26,14 +27,31 @@ ValidPath = Annotated[
     )
 ]
 
+EmbedContext = Annotated[
+    bool,
+    typer.Option(
+        '--embed-context/--no-embed-context'
+    )
+]
+
+PrettyPrint = Annotated[
+    bool,
+    typer.Option(
+        '--pretty-print/--no-pretty-print'
+    )
+]
 
 def parse(path: Path) -> tuple[dict, str]:
     with open(path, 'r') as in_file:
         settings, document = frontmatter.parse(in_file.read())
 
-    # Default graph name.
-    if 'graph-name' not in settings:
-        settings['graph-name'] = path.name
+    if 'id' not in settings and 'title' not in settings:
+        settings['title'] = path.stem
+
+    try:
+        settings['id'] = str(settings['id'])
+    except KeyError:
+        settings['id'] = str(path.name)
 
     return (settings, document)
 
@@ -48,21 +66,29 @@ def configure_logging(debug: bool) -> None:
 
 
 @app.command()
-def transform_and_serialise(path: ValidPath, syntax: str = 'json-ld', debug: bool = False) -> None:
+def transform_and_serialise(path: ValidPath, syntax: str = 'json-ld', debug: bool = False, context: EmbedContext = None, pretty_print: PrettyPrint = True, metadata: bool = False) -> None:
     configure_logging(debug)
     (settings, document) = parse(path)
+    settings['embed-context'] = context
+    settings['metadata'] = metadata
     template = Template(**settings)
     html = template._transform_md_to_html(document)
     doc = html.xpath('/document')[0]
 
-    glossary_path_str = settings.get('glossary')
-    if glossary_path_str:
-        glossary_path = path.parent / glossary_path_str
-        (_, glossary_document) = parse(glossary_path)
-        glossary_html = template._transform_md_to_html(glossary_document)
-        glossary_dls = glossary_html.xpath('/document/dl')
-        for dl in glossary_dls:
-            doc.append(dl)
+    glossaries = settings.get('import')
+    if glossaries:
+        if isinstance(glossaries, str):
+            glossaries = [glossaries]
+        else:
+            glossaries = list(set(glossaries))
+
+        for glossary in glossaries:
+            glossary_path = path.parent / glossary
+            (_, glossary_document) = parse(glossary_path)
+            glossary_html = template._transform_md_to_html(glossary_document)
+            glossary_definition_lists = glossary_html.xpath('/document/dl')
+            for definition_list in glossary_definition_lists:
+                doc.append(definition_list)
 
     (status, result) = template._transform_html_to_json_ld(html)
     if not status:
@@ -85,10 +111,34 @@ def transform_and_serialise(path: ValidPath, syntax: str = 'json-ld', debug: boo
         case _:
             graph = ConjunctiveGraph()
             graph.parse(data=result, format='json-ld')
+            graph.bind('ex', 'http://example.org/')
+            graph.bind('exterms', 'http://example.org/terms/')
+
             try:
                 data = graph.serialize(format=syntax)
-                print(data)
             except PluginException:
                 print(f'Unrecognised RDF syntax: "{syntax}".')
-                print('Try: "nt", "turtle" or "nquads".')
-                typer.Exit(code=1)
+                print('Try: "nt", "turtle" or "nquads".\n')
+                print('See: <https://rdflib.readthedocs.io/en/7.1.1/plugin_serializers.html>')
+                raise typer.Exit(code=1)
+
+            console = Console()
+            if pretty_print and console.is_terminal:
+                syntax_lexer = {
+                    'longturtle': 'turtle',
+                    'n3': 'turtle',
+                    'nquads': 'turtle',
+                    'nt': 'turtle',
+                    'pretty-xml': 'xml',
+                    'trig': 'turtle'
+                }
+                lexer = syntax_lexer.get(syntax, syntax)
+                renderer = Syntax(
+                    data,
+                    lexer,
+                    background_color='default',
+                    word_wrap=True
+                )
+                console.print(renderer)
+            else:
+                print(data)