+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
-## 3.0.0
+## 3.0.0-alpha
 
 - Rename project to Knowledge Representation Markup Language
 - Refactor specification and implementation
 - Add testmark fixtures to the specification
+- Improve CLI
+  - Option: Rich syntax highlighting
+  - Option: embed context document
 
 ## 2.14.4